pyproject.toml
[project]
name = "unified-ui-re-act-agent-service"
version = "0.1.0"
description = "ReACT Agent Service for the unified-ui platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [{ name = "unified-ui", email = "contact@unified-ui.com" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "unifiedui-sdk==0.1.1",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2.1",
    "langchain-ollama>=0.3",
    "langchain-groq>=0.3",
    "langchain-mistralai>=0.3",
    "httpx>=0.27",
    "sse-starlette>=2.0",
    "python-dotenv>=1.0",
    "azure-identity>=1.19",
    "azure-keyvault-secrets>=4.9",
    "hvac>=2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-xdist>=3.5",
    "pytest-asyncio>=0.25",
    "httpx>=0.27",
    "ruff>=0.11",
    "mypy>=1.15",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "TC", "RUF", "D"]
ignore = ["D100", "D104", "D106", "D107"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
